LayoutTests/media/modern-media-controls/time-control/time-control-use-six-digits-for-time-labels.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>TimeControl</code> class property useSixDigitsForTimeLabels.");

const timeControl = new TimeControl;

debug("Testing default value");
shouldBeFalse("timeControl.useSixDigitsForTimeLabels");

debug("");
debug("Setting timeControl.useSixDigitsForTimeLabels = true");
timeControl.useSixDigitsForTimeLabels = true;
shouldBeTrue("timeControl.useSixDigitsForTimeLabels");

timeControl.width = 500;
debug("");
shouldBe("timeControl.elapsedTimeLabel.x", "-2");
shouldBe("timeControl.scrubber.x", "65");
shouldBe("timeControl.scrubber.width", "362");
shouldBe("timeControl.remainingTimeLabel.x", "432");

debug("");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
