LayoutTests/media/modern-media-controls/time-control/time-control-use-six-digits-for-time-labels-expected.txt

Testing the TimeControl class property useSixDigitsForTimeLabels.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing default value
PASS timeControl.useSixDigitsForTimeLabels is false

Setting timeControl.useSixDigitsForTimeLabels = true
PASS timeControl.useSixDigitsForTimeLabels is true

PASS timeControl.elapsedTimeLabel.x is -2
PASS timeControl.scrubber.x is 65
PASS timeControl.scrubber.width is 362
PASS timeControl.remainingTimeLabel.x is 432

PASS successfullyParsed is true

TEST COMPLETE
